sys/ofed: remove all diffs
There is no way this code works and these diffs weren't close enough to
matter.  Drop them until someone is ready for make a concerted effort.

+++ sys/ofed/drivers/infiniband/core/uverbs.h
@@ -244,7 +244,7 @@ int uverbs_dealloc_mw(struct ib_mw *mw);
 void ib_uverbs_detach_umcast(struct ib_qp *qp,
 			     struct ib_uqp_object *uobj);
 
-long ib_uverbs_ioctl(struct file *filp, unsigned int cmd, uintptr_t arg);
+long ib_uverbs_ioctl(struct file *filp, unsigned int cmd, unsigned long arg);
 
 struct ib_uverbs_flow_spec {
 	union {
